Modernise naming and types across repo
- Replace all `double precision` with `real(dp)` everywhere
- Rename dilation variable Dp/D_p -> dilation/dil_u (fixes dp kind collision)
- Rename Get_dD_to_dI -> calc_ddil_by_dI, Get_dD_to_dEpsP -> calc_ddil_by_deps_p
- Invariants modules: all argument/local variable names now lowercase snake_case
- Derivative locals in stress_invariant_deriv.f90: dJ2_dsig -> dJ2_by_dsig,
  II -> ii, TrS2 -> tr_s2
- voigt_utils.f90: calc_two_norm_tensor, calc_two_norm_tensor_strain,
  calc_tensor_inner_product converted from subroutines to pure functions
- Tests: remove random_number, use fixed stress states

+++ src/utils/boolean_helper_funcs.f90
@@ -11,7 +11,7 @@ contains
       ! Boolean
       !******************************************************************
       implicit none
-      double precision, intent(in):: A
+      real(dp), intent(in):: A
       logical, intent(out):: B
       if (A<1.0) then
          B=.false.
